feat: 상품 추가, 수정 API 추가

+++ src/main/java/cart/domain/Money.java
@@ -16,7 +16,7 @@ public class Money {
         this.value = value;
     }
 
-    public static Money of(int value) {
+    public static Money of(long value) {
         if (value < MINIMUM_VALUE) {
             throw new ServiceException(INVALID_MONEY);
         }
